change question to remove assumption
"Why should we not push code when the build is red?" assumes the question "should we not push code when the build is red?" already has an answer of "no", and now just asking why.  But that question itself is valid, and there even are times when we push on red (e.g. the push will fix the redness, some critical situations, etc.)